pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stravalyse"
version = "2.0.0"
description = "A tool to visualise and analyse Strava activities"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Felix van Oost"}]
dependencies = [
    "geojson>=3.1",
    "geopy>=2.4",
    "geopandas>=1.0",
    "matplotlib>=3.9",
    "pandas>=2.2",
    "polyline>=2.0",
    "python-dotenv>=1.0",
    "seaborn>=0.13",
    "stravalib>=2.0",
    "tippecanoe>=2.27",
    "toml>=0.10"]
[project.scripts]
stravalyse = "stravalyse.stravalyse:main"

[tool.pylint.MASTER]
fail-under = 10.0
ignore = 'CVS'
jobs = 0
limit-inference-results = 100
persistent = 'yes'
py-version = '3.12'
suggestion-mode = 'yes'
unsafe-load-any-extension = 'no'

[tool.pylint.REPORTS]
evaluation = '10.0 - ((float(5 * error + warning + refactor + convention) / statement) * 10)'
output-format = 'colorized'
reports = 'yes'
score = 'yes'

[tool.pylint.REFACTORING]
max-nested-blocks = 5
never-returning-functions = 'sys.exit,argparse.parse_error'

[tool.pylint.BASIC]
argument-naming-style = 'snake_case'
attr-naming-style = 'snake_case'
class-attribute-naming-style = 'any'
class-const-naming-style = 'UPPER_CASE'
class-naming-style = 'PascalCase'
const-naming-style = 'UPPER_CASE'
docstring-min-length = -1
function-naming-style = 'snake_case'
include-naming-hint = 'yes'
inlinevar-naming-style = 'any'
method-naming-style = 'snake_case'
module-naming-style = 'snake_case'
no-docstring-rgx = '^_'
property-classes = 'abc.abstractproperty'
variable-naming-style = 'snake_case'

[tool.pylint.FORMAT]
ignore-long-lines = '^\s*(# )?<?https?://\S+>?$'
indent-after-paren = 4
max-line-length = 100
max-module-lines = 1000
single-line-class-stmt = 'no'
single-line-if-stmt = 'no'

[tool.setuptools.packages]
find = {namespaces = false}
